requirements.txt
annotated-types==0.7.0
anyio==4.9.0
Brotli==1.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cssselect2==0.8.0
defusedxml==0.7.1
fastapi==0.115.12
fonttools==4.58.4
fpdf2==2.8.3
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.0
packaging==25.0
pandas==2.3.0
pillow==11.2.1
pluggy==1.6.0
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydyf==0.11.0
Pygments==2.19.1
PyMySQL==1.1.1
pyphen==0.17.2
pytest==8.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
webencodings==0.5.1
websockets==15.0.1
zopfli==0.2.3.post1
